refactor(presentation): split room and garden insight endpoints
