+++ HomeAssistant.Presentation/GardenAdvisor/RoomInsights/Contracts/RoomResponse.cs
